% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDRwate.R
\name{RDRwate}
\alias{RDRwate}
\title{Rate doubly robust (RDR) estimation for weighted average treatment effect (wate)}
\usage{
RDRwate(
  A,
  Y,
  X,
  beta = FALSE,
  v1 = NA,
  v2 = NA,
  method = "EIF",
  n.folds = 5,
  n.split = 10,
  return.naive = FALSE,
  ps.library = c("SL.glm", "SL.glm.interaction"),
  out.library = c("SL.glm", "SL.glm.interaction"),
  seed = 4399
)
}
\arguments{
\item{A}{the treatment vector, which should be valued on 0 or 1, where 0 is the control, and 1 is the treated}

\item{Y}{the outcome vector}

\item{X}{the covariate matrix or vector (if only one covariate)}

\item{beta}{whether to calculate WATEs based on beta weights; if so, specify the `v1` and `v2` parameters}

\item{v1}{a model parameter in beta weights; default is NA (beta weights are not estimated); can be a vector and the length must be the same as v2}

\item{v2}{a model parameter in beta weights; default is NA (beta weights are not estimated); can be a vector and the length must be the same as v1}

\item{method}{method for estimation: "EIF" or "DML";
"EIF" calculates the nuisance functions on the whole dataset, without using cross-fitting;
"DML" uses cross-fitting on k folds (if use DML, the `n.folds` and `n.split` parameters need to be specified)}

\item{n.folds}{number of folds used for cross fitting when using DML methods; default is 5}

\item{n.split}{number of replications for doing the sample splits for cross-fitting when using DML methods; default is 10}

\item{return.naive}{whether to calculate and return the two naive estimators (the propensity score weighted and outcome imputed estimators)}

\item{ps.library}{method(s) used for fitting the propensity score model, chosen from the list in the SuperLearner package, and can be a vector of methods ensemble; see `SuperLearner::listWrappers()`}

\item{out.library}{method(s) used for fitting the outcome regression models, chosen from the list in the SuperLearner package, and can be a vector of methods ensemble; see `SuperLearner::listWrappers()`}

\item{seed}{seed for generating random numbers used in set.seed() function when splitting data; default is 4399}
}
\description{
Rate doubly robust (RDR) estimation for weighted average treatment effect (wate)
}
