% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDRwate.R
\name{RDRwate}
\alias{RDRwate}
\title{Rate doubly robust (RDR) estimation for weighted average treatment effect (WATE)}
\usage{
RDRwate(
  A,
  Y,
  X,
  beta = FALSE,
  v1 = NA,
  v2 = NA,
  method = "EIF",
  n.folds = 5,
  ps.library = c("SL.glm", "SL.glm.interaction"),
  out.library = c("SL.glm", "SL.glm.interaction"),
  seed = 1
)
}
\arguments{
\item{A}{treatment vector; should be 0 or 1 valued, where 0 is control, and 1 is treated}

\item{Y}{outcome vector}

\item{X}{covariate matrix}

\item{beta}{whether to calculate WATE based on beta weights; if so, specify the below v1 & v2 parameters}

\item{v1}{model parameter in beta weights; default is NA (beta weights are not estimated); can be a vector and length must be the same as v2}

\item{v2}{model parameter in beta weights; default is NA (beta weights are not estimated); can be a vector and length must be the same as v1}

\item{method}{method for estimation: "EIF" or "DML";
"EIF" calculates the nuisance functions on the whole dataset, without using cross-fitting;
"DML" uses cross-fitting on k folds (k = n.folds specified below)}

\item{n.folds}{number of folds used for cross fitting; default is 5}

\item{ps.library}{method(s) used for fitting the propensity score model;
it is chosen from the list of the SuperLearner package, and can be a vector of methods (the ensemble library)}

\item{out.library}{method(s) used for fitting the outcome regression models;
it is chosen from the list of the SuperLearner package, and can be a vector of methods (the ensemble library)}

\item{seed}{seed for generating random numbers used in set.seed() function when splitting data; default is 1}
}
\description{
Rate doubly robust (RDR) estimation for weighted average treatment effect (WATE)
}
